+++ Tests/VCoreTests/Helpers/EquatableError.swift
@@ -1,0 +1,11 @@
+//
+//  EquatableError.swift
+//  VCoreTests
+//
+//  Created by Vakhtang Kontridze on 08.06.22.
+//
+
+import Foundation
+
+// MARK: - Equatable Error
+protocol EquatableError: Error, Equatable {}